RL: Remove unneccessary division by zero checks
Linearised expressions for typical gate variables, take the form 1/a,
which cannot be zero, so we can remove the runtime check when dividing
by the linearised expression in the generated code.

It was somewhat tricky to make sympy perform this check without any
substantial increase in the execution time, as sympy supports a number
of expensive checks for the numerator being non-zero. It's better to
make the check inexpensive, and then we can fall back to generating the
extra runtime check, even if it may be possible to prove that division
by zero cannot possibly occur.

The final implementation does not seem to increase the execution time
of gotran, and it should lead to slightly faster execution time in the
generated code as well.